tests/requirements.txt
Pillow
git+git://github.com/zmwangx/pyzmwangx.git@master
coverage
nose
